feat(testa gen): rename service
