+++ backend/src/lastwill/lastwill.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common'
 import { DbModule } from '../db/db.module'
 import { SharedModule } from '../shared/shared.module'
 import { LastWillController } from './lastwill.controller'
-import { LastWillService } from './services/lastwill.service'
+import { LastWillService } from './lastwill.service'
 
 @Module({
   imports: [SharedModule, DbModule],
